refactor forward browser logs
