+++ packages/next/src/next-devtools/userspace/app/app-dev-overlay-setup.ts
@@ -1,13 +1,8 @@
 import { patchConsoleError } from './errors/intercept-console-error'
 import { handleGlobalErrors } from './errors/use-error-handler'
-import {
-  initializeDebugLogForwarding,
-  isTerminalLoggingEnabled,
-} from './forward-logs'
+import { initializeDebugLogForwarding } from './forward-logs'
 
 handleGlobalErrors()
 patchConsoleError()
 
-if (isTerminalLoggingEnabled) {
-  initializeDebugLogForwarding('app')
-}
+initializeDebugLogForwarding('app')
