+++ packages/next/src/next-devtools/userspace/pages/pages-dev-overlay-setup.tsx
@@ -15,7 +15,6 @@ import {
   forwardUnhandledError,
   logUnhandledRejection,
   forwardErrorLog,
-  isTerminalLoggingEnabled,
 } from '../app/forward-logs'
 
 const usePagesDevOverlayBridge = () => {
@@ -82,9 +81,7 @@ function nextJsHandleConsoleError(...args: any[]) {
   storeHydrationErrorStateFromConsoleArgs(...args)
   // TODO: Surfaces non-errors logged via `console.error`.
   handleError(maybeError)
-  if (isTerminalLoggingEnabled) {
-    forwardErrorLog(args)
-  }
+  forwardErrorLog(args)
   origConsoleError.apply(window.console, args)
 }
 
@@ -92,7 +89,7 @@ function onUnhandledError(event: ErrorEvent) {
   const error = event?.error
   handleError(error)
 
-  if (error && isTerminalLoggingEnabled) {
+  if (error) {
     forwardUnhandledError(error as Error)
   }
 }
@@ -109,9 +106,7 @@ function onUnhandledRejection(ev: PromiseRejectionEvent) {
   }
 
   dispatcher.onUnhandledRejection(reason)
-  if (isTerminalLoggingEnabled) {
-    logUnhandledRejection(reason)
-  }
+  logUnhandledRejection(reason)
 }
 
 export function register() {
@@ -124,9 +119,7 @@ export function register() {
     Error.stackTraceLimit = 50
   } catch {}
 
-  if (isTerminalLoggingEnabled) {
-    initializeDebugLogForwarding('pages')
-  }
+  initializeDebugLogForwarding('pages')
   window.addEventListener('error', onUnhandledError)
   window.addEventListener('unhandledrejection', onUnhandledRejection)
   window.console.error = nextJsHandleConsoleError
